+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,32 +1,12 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findById(long id);
 
-    public UserService(UserDao userDao, UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(User user);
 
-    public User findById(long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(long id, String newPassword, String createBy) {
-        User user = findById(id);
-        user.changePassword(newPassword);
-        userDao.update(user);
-        userHistoryDao.log(new UserHistory(user, createBy));
-    }
+    void changePassword(long id, String newPassword, String createBy);
 }
